Fix when PaymentMethod it's prefered, buttons (star & trash) hidden

+++ resources/views/livewire/payment-methods.blade.php
@@ -80,15 +80,20 @@
                                 <p>Expira: {{ $paymentMethod->card->exp_month }}/{{ $paymentMethod->card->exp_year }}</p>
                             </div>
 
-                            <div class="flex space-x-4 mr-4">
-                                <button class="disabled:opacity-25" wire:click="defaultPaymentMethod('{{ $paymentMethod->id }}')" wire:target="defaultPaymentMethod('{{ $paymentMethod->id }}')" wire:loading.attr="disabled">
-                                    <i class="fa-regular fa-star"></i>
-                                </button>
-    
-                                <button class="disabled:opacity-25" wire:click="deletePaymentMethod('{{ $paymentMethod->id }}')" wire:target="deletePaymentMethod('{{ $paymentMethod->id }}')" wire:loading.attr="disabled">
-                                    <i class="fa-regular fa-trash-can"></i>
-                                </button>
-                            </div>
+                            @if ($this->defaultPaymentMethod->id !== $paymentMethod->id)
+                                
+                                <div class="flex space-x-4 mr-4">
+                                    <button class="disabled:opacity-25" wire:click="defaultPaymentMethod('{{ $paymentMethod->id }}')" wire:target="defaultPaymentMethod('{{ $paymentMethod->id }}')" wire:loading.attr="disabled">
+                                        <i class="fa-regular fa-star"></i>
+                                    </button>
+        
+                                    <button class="disabled:opacity-25" wire:click="deletePaymentMethod('{{ $paymentMethod->id }}')" wire:target="deletePaymentMethod('{{ $paymentMethod->id }}')" wire:loading.attr="disabled">
+                                        <i class="fa-regular fa-trash-can"></i>
+                                    </button>
+                                </div>
+
+                            @endif
+
                         </li>
 
                     @endforeach
